Update Prob3.cpp
Hi Shaurya Bhaiyya ,
I am Pranay and While practising this problem I found a little bug in it. The problem statement clearly says that you can buy at most one box per day. So this code is failing for the test case N=1, D=2, and E=10; So I made the changes accordingly.
Plz take this pull request cross verify and then  update the code.
Also in last while calculating the req for days if the left sweets are greater as compared to the req of days then just return ans;
Ok Thank you